Tab width is now configurable. Added draw_char function in display.c which splits out some functionality from draw_line.

+++ variable.h
@@ -32,7 +32,7 @@ typedef enum {
 typedef struct {
     ValueType type;
     union {
-        int ival;
+        long ival;
         double fval;
         char *sval;
     } val;
